extract point editor
